build: 🧪 fix build 📦
