+++ packages/blog/gatsby/createBlogPostPages.ts
@@ -1,0 +1,57 @@
+import { CreatePagesArgs } from "gatsby"
+import path from "path"
+import { CreateBlogPostPagesQuery } from "../../src/types/generated"
+
+export const createBlogPostPages = async ({
+  graphql,
+  actions,
+  reporter,
+}: CreatePagesArgs) => {
+  const BlogPostTemplate = path.resolve(`src/templates/BlogPostTemplate.tsx`)
+
+  const result = await graphql<CreateBlogPostPagesQuery>(`
+    query CreatePageQuery {
+      allMarkdownRemark(
+        sort: { fields: [frontmatter___date], order: ASC }
+        limit: 1000
+      ) {
+        nodes {
+          id
+          fields {
+            slug
+          }
+        }
+      }
+    }
+  `)
+
+  // Handle errors
+  if (result.errors) {
+    reporter.panicOnBuild(
+      `There was an error loading your blog posts`,
+      result.errors
+    )
+    return
+  }
+
+  const posts = result.data.allMarkdownRemark.nodes
+
+  if (posts.length > 0) {
+    const { createPage } = actions
+
+    posts.forEach((post, index) => {
+      const previousPostId = index === 0 ? null : posts[index - 1].id
+      const nextPostId = index === posts.length - 1 ? null : posts[index + 1].id
+
+      createPage({
+        path: post.fields.slug,
+        component: BlogPostTemplate,
+        context: {
+          id: post.id,
+          previousPostId,
+          nextPostId,
+        },
+      })
+    })
+  }
+}
