+++ packages/blog/gatsby/gatsby-browser.ts
@@ -2,9 +2,9 @@
 import "typeface-montserrat"
 import "typeface-merriweather"
 // normalize CSS across browsers
-import "./src/normalize.css"
+import "../src/normalize.css"
 // custom CSS styles
-import "./src/style.css"
+import "../src/style.css"
 
 // Highlighting for code blocks
 import "prismjs/themes/prism.css"
