--- packages/blog/src/pages/queries/NotFoundQuery.d.ts
@@ -1,20 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-// @generated
-// This file was automatically generated and should not be edited.
-
-// ====================================================
-// GraphQL query operation: NotFoundQuery
-// ====================================================
-
-export interface NotFoundQuery_site_siteMetadata {
-  title: string | null;
-}
-
-export interface NotFoundQuery_site {
-  siteMetadata: NotFoundQuery_site_siteMetadata | null;
-}
-
-export interface NotFoundQuery {
-  site: NotFoundQuery_site | null;
-}
